refactor(migration): separate how a request reaches a chain from what it asks
Transports, providers, snapshots, impersonation, balance writes and the
fee-history compatibility shims now live in their own module. They are
about *how* a request reaches a node rather than what the migration wants
from it, and that is where a rehearsal and a live run differ almost
entirely — impersonation instead of keys, state controls instead of
waiting. Keeping the two apart stops that difference leaking into the
phase logic.

`impersonationProvider` stays behind: it takes the deploy option bag, so
moving it would have dragged the deploy path along with it.

`migration.ts` drops from 10,332 to 9,923 lines.
